improved frontend app startup and user registration process
