+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/@core/components/nav-shell/nav-shell.component.ts
@@ -3,7 +3,7 @@ import { ShellIconInfo } from '../../../../models/ShellIconInfo';
 import { FriendRequestDTO } from '../../../../models/FriendRequestDTO';
 import { FriendRequestService } from '../../services/friendrequest.service';
 import { ShellInteractionService } from '../../services/shell-interaction.service';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../services/auth/auth.service';
 import { Component, ViewChild, OnInit, OnDestroy } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { BehaviorSubject, Subscription } from 'rxjs';
