+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/@core/components/settings/settings.component.ts
@@ -1,6 +1,6 @@
 import { CreateOrUpdateUserDTO } from './../../../../models/CreateOrUpdateUserDTO';
 import { UserDTO } from '../../../../models/UserDTO';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../services/auth/auth.service';
 import { UserService } from 'src/app/@core/services/user.service';
 import { SettingsService } from '../../services/settings.service';
 import { Component, OnDestroy, OnInit } from '@angular/core';
