+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/@core/services/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from './auth/auth.service';
 
 @Injectable({providedIn: 'root'})
 export class AuthGuard implements CanActivate {
