+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/@core/services/auth/ProfileInfoResult.ts
@@ -1,0 +1,9 @@
+import { UserProfile } from 'src/models/UserProfile';
+
+export class ProfileInfoResult {
+
+  constructor(
+    public profile: UserProfile,
+    public isRegistered: boolean) {
+  }
+}
