+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/invitation/components/welcome-invited/welcome-invited.component.html
@@ -11,7 +11,7 @@
         <span>!</span>
       </div>
 
-      <img class="pulsate-transform" (click)="onLogin()" src="../../assets/img/Logo.png">
+      <img class="pulsate-transform" (click)="onLogin()" src="./../../../../assets/img/logo.png">
     </div>
   </div>
