+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/onboarding/components/onboarding/onboarding.component.scss
@@ -15,3 +15,7 @@ img {
 button.right {
     align-self: flex-end;
 }
+
+.margin-right {
+  margin-right: 10px;
+}
