+++ BingeBuddyNg/BingeBuddyNg.Spa/src/app/onboarding/components/welcome/welcome.component.ts
@@ -1,23 +1,36 @@
-import { Router } from '@angular/router';
-import { AuthService } from '../../../@core/services/auth.service';
-import { Component, OnInit } from '@angular/core';
+import { Router, ActivatedRoute } from '@angular/router';
+import { AuthService } from '../../../@core/services/auth/auth.service';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { filter } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-welcome',
   templateUrl: './welcome.component.html',
   styleUrls: ['./welcome.component.css']
 })
-export class WelcomeComponent implements OnInit {
+export class WelcomeComponent implements OnInit, OnDestroy {
+
+  private sub: Subscription;
 
   constructor(public auth: AuthService,
     private router: Router) { }
 
   ngOnInit() {
-    if (this.auth.isAuthenticated()) {
-      this.router.navigate(['/activity-feed']);
-    }
+    console.log('WelcomeComponent: ngOnInit');
+    this.sub = this.auth.currentUserProfile$
+      .pipe(filter(p => p != null))
+      .subscribe(() => {
+        console.log('WelcomeComponent: Navigating');
+        this.router.navigate(['/activity-feed']);
+      });
   }
 
+  ngOnDestroy() {
+    if (this.sub) {
+      this.sub.unsubscribe();
+    }
+  }
 
   onLogin() {
     this.auth.login();
